+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-19T14:53:38.511Z",
+    "2026-06-20T23:11:59.550Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {