+++ packages/orq-rc/RELEASES.md
@@ -10139,3 +10139,13 @@ Based on:
 - [typescript v4.11.0-rc.35] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.35] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.35 - packages/orq-rc
+
+## 2026-06-20 23:19:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.784.0 (2.911.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.36] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.36 - packages/orq-rc